fixed reverse order of ints issues still in description
Only works when starting with w140 for some reason.
Only shows up to 7 things at a time which isn't a problem, more of an issue because I am unsure why.
Sword of Edicts no matter how I try to call it does not work and is ignored.
Lots of test spam.
Probably more.