Deleted to-do.md to make everyone else happy.
Fixed issue #5 because apparently this to-do list is stupid and silly and didn't need to be here. Regardless of my opinion on it, I decided to comply to reduce the stress between this time. If I had my choice we would have 3 to-do lists all filled, but we don't, we had one. Soon to have none. Hopefully, we don't get un-organized without this to-do list and spiral into chaos for the lack of next steps and guidance. If this team falls apart and we cannot finish this bot to be the best it could ever be, I put this blame on you Leon. This will be entirely your fault for your opinions and preferences and bias towards GitHub issues and lack of willfulness to try something new and love the checkboxes that I feel like I am being forced to be rid of. Anyhow- consider this to-do list removed, may it not upset you too much.

-[x] remove to-do list.